Added multi_run.sh to help with the gaussian functions.
With the help of this program it is easy to run multiple files for the execution all just at once.
It uses programs such as nice --15 which sets the priority level to -15 which makes it faster than running with normal priorities.
Not only this it also uses nohup command which makes the process to run even in the background as well as after closing the terminal session,
as long as the process execution is not finished and the server is up and running.